vr1.2
updated the side borders to remove the right one as im not sure what i want to do there.  modified the left one to prepare to move the site nav there and make a page nav at the top.  I also attempted to remove the overlays from the map but that will require using google maps api and i am not at the stage to integrate backend dev but i did align the map better.  I also adjusted some headings.  Next update will include the moving of the navs and spreading these changes to the other pages for now.  The nav will include some javascript for changing the size of the site nav, which will include icons and the js will expand to include the words and shrink to just show the icons